major refactor
* user interface is the same though
