+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dylink"
-version = "0.5.2"
+version = "0.5.3"
 authors = ["Jonathan Thomason"]
 edition = "2021"
 license = "MIT OR Apache-2.0"
@@ -16,4 +16,3 @@ members = ["dylink_macro"]
 [dependencies.dylink_macro]
 version = "0.6"
 path="./dylink_macro"
-features = []
